perf: switch vector index to float32 — halves memory, eliminates conversions
Embeddings arrive as float32 from providers and are stored as float32
blobs in SQLite. The vector index was promoting them to float64 for
computation — doubling memory with zero precision benefit at 768-dim.

- VectorEntry.Embedding: []float64 → []float32
- VectorIndex.Search() query: []float64 → []float32
- Add CosineSimilarityF32() alongside existing F64 (accumulates in float64)
- BuildFromStore() no longer converts float32→float64 on load
- embedAndStore() passes vec directly, no conversion loop
- HybridSearch() passes queryEmbedding directly to Search()

Memory per 768-dim entry: ~6KB → ~3KB. At 50K entries: ~300MB → ~150MB.

+++ internal/agent/memory/retrieval.go
@@ -72,7 +72,7 @@ type MemoryEntry struct {
 // RetrieveWithANN uses the vector index for approximate nearest-neighbor
 // search over memory embeddings. Falls back to empty results if the index is
 // not built or has insufficient entries.
-func (mr *Retriever) RetrieveWithANN(ctx context.Context, embedding []float64, k int) []MemoryEntry {
+func (mr *Retriever) RetrieveWithANN(ctx context.Context, embedding []float32, k int) []MemoryEntry {
 	if mr.vectorIndex == nil || !mr.vectorIndex.IsBuilt() || k <= 0 {
 		return nil
 	}
